- Improved readiness of Readings processing background job - Improved Background job registration and setup, - Minor code cleaning in Infrastructure project
