+++ src/SmartAc.Infrastructure/AlertResolvers/MonoxideInRange.cs
@@ -3,7 +3,7 @@ using SmartAc.Application.Options;
 using SmartAc.Domain.Alerts;
 using SmartAc.Domain.DeviceReadings;
 
-namespace SmartAc.Infrastructure.Resolvers;
+namespace SmartAc.Infrastructure.AlertResolvers;
 
 internal sealed class MonoxideInRange : Resolver
 {
