+++ src/SmartAc.Infrastructure/AlertResolvers/Resolver.cs
@@ -2,7 +2,7 @@
 using SmartAc.Domain.Alerts;
 using SmartAc.Domain.DeviceReadings;
 
-namespace SmartAc.Infrastructure.Resolvers;
+namespace SmartAc.Infrastructure.AlertResolvers;
 
 internal class Resolver
 {
