+++ src/SmartAc.Infrastructure/AlertResolvers/TempInRange.cs
@@ -3,12 +3,12 @@ using SmartAc.Application.Options;
 using SmartAc.Domain.Alerts;
 using SmartAc.Domain.DeviceReadings;
 
-namespace SmartAc.Infrastructure.Resolvers;
+namespace SmartAc.Infrastructure.AlertResolvers;
 
-internal sealed class HumidityInRange : Resolver
+internal sealed class TempInRange : Resolver
 {
     public override bool IsResolved(DeviceReading reading, AlertType alertType, SensorParams sensorParams)
-        => alertType == AlertType.OutOfRangeHumidity && 
-           reading.Humidity.InRange(sensorParams.HumidityPctMin, sensorParams.HumidityPctMax) || 
+        => alertType == AlertType.OutOfRangeTemp &&
+           reading.Temperature.InRange(sensorParams.TemperatureMin, sensorParams.TemperatureMax) ||
            base.IsResolved(reading, alertType, sensorParams);
 }
