+++ src/SmartAc.Infrastructure/Extensions/ServiceCollectionExtensions.cs
@@ -3,7 +3,6 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 using Quartz;
 using SmartAc.Application.Abstractions.Services;
-using SmartAc.Infrastructure.BackgroundJobs;
 using SmartAc.Infrastructure.Options;
 using SmartAc.Infrastructure.Services;
 
@@ -13,8 +12,8 @@ public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
     {
+        RegisterServices(services, configuration);
         RegisterBackgroundServices(services, configuration);
-        RegisterJwtService(services);
         RegisterOptions(services, configuration);
 
         return services;
@@ -31,30 +30,12 @@ public static class ServiceCollectionExtensions
 
     private static void RegisterBackgroundServices(IServiceCollection services, IConfiguration configuration)
     {
-        services.AddQuartz(configure =>
-        {
-            var jobKey = new JobKey(nameof(ProcessNewReadingsJob));
-
-            var intervalInSeconds =
-                configuration
-                    .GetRequiredSection("BackgroundJobParams")
-                    .GetValue<int>("IntervalInSeconds");
-
-            configure
-                .AddJob<ProcessNewReadingsJob>(jobKey)
-                .AddTrigger(trigger =>
-                    trigger.ForJob(jobKey)
-                        .WithSimpleSchedule(schedule =>
-                            schedule.WithIntervalInSeconds(intervalInSeconds)
-                                .RepeatForever()));
-
-            configure.UseMicrosoftDependencyInjectionJobFactory();
-        });
-
-        services.AddQuartzHostedService();
+        services.AddQuartz(configure => configure.UseMicrosoftDependencyInjectionJobFactory());
+        services.AddQuartzHostedService(options => options.WaitForJobsToComplete = true);
+        services.ConfigureOptions<ReadingProcessingJobSetup>();
     }
 
-    private static void RegisterJwtService(IServiceCollection services)
+    private static void RegisterServices(IServiceCollection services, IConfiguration configuration)
     {
         services.TryAddTransient<ISmartAcJwtService, SmartAcJwtService>();
     }
